First project settings.
