+++ BruteForce/Resources/Assets.xcassets/Contents.json
