+++ BruteForce/Sources/Controllers/PasswordViewController.swift
@@ -1,5 +1,5 @@
 //
-//  ViewController.swift
+//  PasswordViewController.swift
 //  BruteForce
 //
 //  Created by Arthur on 17.10.2022.
@@ -7,13 +7,9 @@
 
 import UIKit
 
-class ViewController: UIViewController {
+class PasswordViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        // Do any additional setup after loading the view.
     }
-
-
 }
-
