--- BruteForceUITests/BruteForceUITestsLaunchTests.swift
@@ -1,32 +1,0 @@
-//
-//  BruteForceUITestsLaunchTests.swift
-//  BruteForceUITests
-//
-//  Created by Arthur on 17.10.2022.
-//
-
-import XCTest
-
-final class BruteForceUITestsLaunchTests: XCTestCase {
-
-    override class var runsForEachTargetApplicationUIConfiguration: Bool {
-        true
-    }
-
-    override func setUpWithError() throws {
-        continueAfterFailure = false
-    }
-
-    func testLaunch() throws {
-        let app = XCUIApplication()
-        app.launch()
-
-        // Insert steps here to perform after app launch but before taking a screenshot,
-        // such as logging into a test account or navigating somewhere in the app
-
-        let attachment = XCTAttachment(screenshot: app.screenshot())
-        attachment.name = "Launch Screen"
-        attachment.lifetime = .keepAlways
-        add(attachment)
-    }
-}
